Eulers.py

+++ Eulers.py
@@ -3,7 +3,7 @@ x=float(input("enter x0 :"))
 y=float(input("enter y0 :"))
 h=float(input("enter h :"))
 fun=input("enter function[NOTE:USE * FOR MULTIPLICATION,USE / FOR DIVISION] :")
-n1=int(input("enter value of x :"))
+n1=float(input("enter value of x :"))
 n2=n1//h
 n=int(n2)
 print("Number of Iterations :",n)
